+++ backend/core/assets.py
@@ -98,16 +98,15 @@ class Assets(object):
             return asset
 
         # 如果是debug环境，使用request中的host来替换localhost
-        # assets_url = self.assets_url
-        #
-        # if self.app.debug and self.assets_debug_url:
-        #     host = request.host.split(':')[0]
-        #     assets_url = self.assets_debug_url.format(host)
+        assets_url = self.assets_url
+
+        if self.app.debug and self.assets_debug_url:
+            host = request.host.split(':')[0]
+            assets_url = self.assets_debug_url.format(host)
 
         for key in self.assets:
             if key == asset:
-                # return '{0}{1}'.format(assets_url, self.assets[key])
-                return self.assets[key]
+                return '{0}{1}'.format(assets_url, self.assets[key])
 
         # 如没有对应的资源文件，则返回传入值以方便查找错误的文件
         return asset