+++ frontend/artui/react/index.js
@@ -2,12 +2,15 @@
  * 对Semantic UI的基于React的扩展.
 **/
 
+import 'react-timeout';
+
 export Hoverable from './behaviors/hoverable';
 
 export Message from './collections/message';
 
 export {Link, IndexLink} from './elements/link';
 export Input from './elements/input';
+export Label from './elements/label';
 
 export Popup from './modules/popup';
 export Tab from './modules/tab';